refactor: with tsx
